New Feature: Compute the cursor position  (#134)
In essence, the line is calculated by dividing the y-position of the
text segment with the cursor by the line height:
```swift 
var line = Int(textSegmentFrame.maxY / textSegmentFrame.height)
```
However, this counts the preceding line wraps as lines too. As a result,
I have to count the preceding line wraps with:
```swift
textLayoutManager.enumerateTextLayoutFragments(from: textLayoutManager.documentRange.location,
                                                           options: [.ensuresLayout, .ensuresExtraLineFragment]
            ) { textLayoutFragment in

                guard let cursorTextLineFragment = textLayoutManager.textLineFragment(at: insertionPointLocation)
                else { return false }

                /// Check whether the textLayoutFragment has line wraps
                if textLayoutFragment.textLineFragments.count > 1 {
                    for lineFragment in textLayoutFragment.textLineFragments {
                        lineWrapsCount += 1
                        /// Do not count lineFragments after the lineFragment where the cursor is placed
                        if lineFragment == cursorTextLineFragment { break }
                    }

                    /// The first lineFragment will be counted as an actual line
                        lineWrapsCount -= 1
                }

                if textLayoutFragment.textLineFragments.contains(cursorTextLineFragment) {
                    return false
                }
                return true
            }
```
Unfortunately, this does scale with the line count of the file. So we
might want to change that if we can come up with a better alternative in
the future. As a first implementation, I think it works.

---------

Co-authored-by: Lukas Pistrol <l.pistrol@gmail.com>