Add Letter Spacing (#174)
### Description

Adds a `letterSpacing` parameter that determines the amount of space
between characters. The parameter is a multiplier of the font's
character width so `1.0` indicates no space, `2.0` indicates a full
character width's space. This is exactly as described in #153.

In detail:
- The letter spacing parameter modifies a stored `kern` variable that is
used for all default attributes.

This PR also adds documentation for a couple undocumented parameters in
the initializer and moves `lineHeight` into the `STTextViewController`
initializer for a cleaner `CodeEditTextView`.

### Related Issues

* closes #153 

### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

The below recording shows a placeholder settings adjusting the letter
spacing, as well as the spacing being applied while typing.


https://user-images.githubusercontent.com/35942988/230160550-c540523c-4fb9-4984-8564-5bdebb7fbcb3.mov