Toggle comment for current line via CMD + / (#241)
This change implements the functionality of <kbd>⌘</kbd> <kbd>/</kbd>
for single line comments. It allows you to toggle between commented and
uncommented for the line the cursor is currently on when you press
<kbd>⌘</kbd> <kbd>/</kbd>.

To do so, I implemented a `keyDown` event recognizer, which listens for
when the relevant keys are pressed. If <kbd>⌘</kbd> <kbd>/</kbd> is
pressed, it calls a method called `commandSlashCalled()`, which decides
which toggle is supposed to happen depending on if the line is already
commented or not. It also addresses the situation of special cases of
languages like HTML, which need a comment at the beginning and end of
the line (essentially a `rangeComment`) to comment a single line.

### Related Issues

- #38

This PR accomplishes part of #38. I talked with some of the project
leads (@fastestMolasses) and they said it makes sense to break #38 into
a couple different issues (I.e. single-line vs highlighted chunks, etc).
Single-line comment toggling is completed as of this PR but commenting
highlighted code still needs to be completed.

### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

<img width="626" alt="Screenshot 2024-04-19 at 5 44 24 PM"
src="https://github.com/CodeEditApp/CodeEditSourceEditor/assets/143217945/1200a5c1-d8ea-48a4-9704-70db7aa23fc7">

---------

Co-authored-by: Abe <abe.malla8@gmail.com>