Multiple Highlighter Support (#273)
### Description

Adds support for multiple highlight providers.

> [!NOTE]
> For reviewers: You may notice that this uses an underscored module
`_RopeModule`. This module is safe, as in it has tests and is used in
production (it backs AttributedString and BigString Foundation types).
It's underscored because the API may change in the future, and the
swift-collections devs consider it to be the incorrect collection type
to use for most applications. However, this application meets every
requirement for using a Rope.

### Related Issues

* #40

### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

N/A

---------

Co-authored-by: Tom Ludwig <tommludwig@icloud.com>