Move Overscroll To TextView (#296)
### Description

The current overscroll behavior that was being handled in the
SourceEditor relied on a content inset, which was a little buggy and did
not allow the user to click in the over scrolled area to set the cursor.
This behavior was moved to TextView instead in order to fix this.

### Related Issues


### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

---------

Co-authored-by: Khan Winter <35942988+thecoolwinter@users.noreply.github.com>