Highlighter Highlights Text in Minimap (#308)
### Description

The highlighter should now highlight all ‘visible’ text, including text that is visible in the minimap but not in the TextView. This PR merges the visible text in the minimap with the visible text in the TextView to highlight both. If the minimap is hidden, the highlighter will behave as before and only highlight the text visible in the TextView.

### Related Issues

* #33 

### Checklist

- [x] I read and understood the [contributing guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md) as well as the [code of conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

https://github.com/user-attachments/assets/4bb7ead0-947d-41bc-96c1-7c7c4df3ad8b