Find and Replace Refactor (#311)
### Description

@austincondiff has been building out the 'replace' functionality and
realized the way we had structured the panel, controller, target, and
model was extremely overcomplicating things. This PR is an attempt to
fix that

Changes:
- Moves all 'business logic' into the `FindPanelViewModel` observable
object. This includes clarified methods like `find`, `replace`, and
`moveToNextMatch/moveToPreviousMatch`. All state has been moved to this
object as well, out of a combination of both the SwiftUI view and the
find view controller.
- Removes the `FindPanelDelegate` type entirely. All that type was doing
was passing method calls from the find panel to it's controller. Since
all that logic is now in the shared view model, the controller & view
can just call the necessary methods on the model.
- Simplifies the `FindViewController` to only handle view/model setup
and layout.
- Changes the focus state variable to an enum instead of two `Bool`s.
This fixes an issue where there was a moment of nothing being focused
when switching between the find and replace text fields.
- Removes the unnecessary `NSHostingView -> NSView -> SwiftUI View`
structure, replacing it with an `NSHostingView` subclass
`FindPanelHostingView` that hosts a `FindPanelView`.
- Clarifies some view naming to reflect what each type does.
  - `FindPanel` -> `FindPanelHostingView`
  - `FindPanelView` search fields moved to:
    - `FindSearchField`
    - `ReplaceSearchField`

### Related Issues

* #295 

### Checklist

<!--- Add things that are not yet implemented above -->

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots

<!--- REQUIRED: if issue is UI related -->

<!--- IMPORTANT: Fill out all required fields. Otherwise we might close
this PR temporarily -->