Code Folding (#341)
### Description

Adds a code folding ribbon.

> [!NOTE]
> For reviewers: Most of these changes have been reviewed individually
in separate PRs. This shouldn't require a detailed combing over. If you
can review a few things in particular like: architecture, naming, and
bugs with real use I think those would be the best use of your limited
time.

#### Detailed Changes
- Folding-related changes
- Added `FoldingRibbonView` - A view that draws a set of folding ranges.
This does all the drawing manually to avoid having to manage view reuse,
it does the drawing in the regions macOS designates so it only draws a
small part of the visible ribbon at a time.
- Added `LineFoldProvider` - A protocol for classes to provide fold
information to CodeEditSourceEditor. This takes a line number, range,
depth, and a reference to the text controller and returns a list of
'line info' objects.
- Added `LineIndentationFoldProvider`, an implementation of
`LineFoldProvider` that calculates folds based on the user's indent
setting.
- Added `LineFoldCalculator`, which interfaces with the given
`LineFoldProvider` to calculate new folds. This happens *asynchronously*
and recalculates every so often. It's very fast but stays off the main
thread to avoid ever gumming up the user's typing.
- Added `LineFoldingModel`, which is the glue between the text view,
calculator, and view.
- To display folded folds, we display a `LineFoldPlaceholder` in the
text view. This keeps a reference to the `LineFoldingModel` via a
delegate protocol to tell the model when it's dismissed.

<details><summary>This is a slightly complicated object graph, so I've
added a diagram here</summary>

```mermaid
flowchart TD
    TV["TextView"] -->|Text Updates| LFM["LineFoldingModel"]
    PL["LineFoldPlaceholder"] --> |Placeholder Removed| LFM
    LFM -->|Text Updates| LFC["LineFoldCalculator"]
    LFC -->|LineFoldStorage| LFM
    LFM -->|Storage Updated| FRV["FoldingRibbonView"]
    LFC <--> |Fold Info| LFP["any LineFoldProvider"]
```

</details>

- Gutter View changes
- New property `showFoldingRibbon` toggles the visibility of the folding
ribbon view.
  - The gutter now manages the folding ribbon view as a subview.
- Added an additional padding option on the gutter for the folding
ribbon.
- Makes a slight modification to how it draws line numbers by using a
line iterator in the drawn lines, rather than the 'visible lines'. Uses
less resources when macOS only requests a redraw for a specific rect
rather than the whole gutter.
- Other, related changes
- Modified the minimap. The minimap is now an attachment delegate to the
text view's attachment object. It receives updates to the text view's
attachments and adds a fake attachment to it's own layout manager. This
effectively syncs the text view's attachments to the minimap.
- Renamed `DispatchQueue.syncMainIfNot` to `waitMainIfNot` and updated
it to use a better method for waiting for a dispatched work item.
  - Added an extension to `NSBezierPath` for rounding corners.
- Added a new `NSColor` convenience initializer for creating light and
dark mode colors more easily.
- Moved the font character width calculation from `TextViewController`
to an `NSFont` extension.

### Related Issues

* #43 

### Checklist

- [x] Folds reflected in minimap.
- [x] Double-click fold expands it.
- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots


https://github.com/user-attachments/assets/25e46ae4-3db5-4f08-8027-1fea9a8e64cb

---------

Co-authored-by: Austin Condiff <austin.condiff@gmail.com>