Jump To Definition (#348)
### Description

Adds jump to definition, including a new key command (control-command-J)
to initiate a jump, and an interaction (command + mouse click). The
destination(s) for a jump is provided via a new object protocol,
provided via the SwiftUI API. The popover presented when multiple links
are present reuses the completion window, with a slight modification so
that it appears as a popover with an arrow rather than a floating
window.

Detailed changes:
- Numerous fixes and improvements to the `SuggestionController` and
`SuggestionViewController`, making them work when presented as a
popover, as well as some resizing bugs.
- New `CodeSuggestionPreviewView`. Displays preview information for a
code suggestion, including a syntax-highlighted source preview,
documentation, and a path component.
- This only appears when one of the three potential information is
available, otherwise it is hidden.
- This view is a visual effect view, so it will inherit the style of the
suggestion controller from the user's theme with a slight tint.
- New `JumpToDefinitionModel` manages the mouse interaction, as well as
providing the layer between the jump delegate and the code suggestion
window.
- New `JumpToDefinitionDelegate` protocol defines how the editor can use
the delegate object to query links and open them once selected.

### Related Issues

* https://github.com/CodeEditApp/CodeEdit/issues/1902

### Checklist

- [x] I read and understood the [contributing
guide](https://github.com/CodeEditApp/CodeEdit/blob/main/CONTRIBUTING.md)
as well as the [code of
conduct](https://github.com/CodeEditApp/CodeEdit/blob/main/CODE_OF_CONDUCT.md)
- [x] The issues this PR addresses are related to each other
- [x] My changes generate no new warnings
- [x] My code builds and runs on my machine
- [x] My changes are all related to the related issue above
- [x] I documented my code

### Screenshots



https://github.com/user-attachments/assets/81d5b24c-6332-4a33-8d2a-8f637bec3bce

---------

Co-authored-by: Abe M <abe.malla8@gmail.com>
Co-authored-by: Wesley de Groot <email@wesleydegroot.nl>